requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-WTF==1.2.2
WTForms==3.2.1
alembic==1.14.1
argon2-cffi==23.1.0
bcrypt==4.2.1
Flask-Limiter==3.10.1
python-dotenv==1.0.1
email-validator==2.2.0
psycopg[binary]==3.2.3
openpyxl==3.1.5
reportlab==4.2.5
pypdf==5.1.0
Pillow==11.1.0
gunicorn==23.0.0
